Natural_Forest/requirements.txt
asgiref==3.8.1
# Django==4.2.3
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
